Also build tests with cmake
