+++ assembler/d16-lib/CMakeLists.txt
@@ -1,10 +1,8 @@
-cmake_minimum_required (VERSION 3.0)
-project (d16-asm)
+project (d16-lib)
 
 set (d16asm_VERSION_MAJOR 1)
 set (d16asm_VERSION_MINOR 0)
 
-list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_LIST_DIR}/cmake")
 find_package(BISON REQUIRED)
 find_package(FLEX REQUIRED)
 find_package(Glib REQUIRED)
@@ -18,15 +16,23 @@ configure_file (
 
 # add the binary tree to the search path for include files
 # so that we will find d16asm_config.h and bison/flex outputs
-include_directories("${PROJECT_BINARY_DIR}" "${PROJECT_SOURCE_DIR}")
-include_directories("${Glib_INCLUDE_DIR}" "${GlibConfig_INCLUDE_DIR}")
+include_directories(${PROJECT_BINARY_DIR} ${PROJECT_SOURCE_DIR})
+include_directories(${Glib_INCLUDE_DIR} ${GlibConfig_INCLUDE_DIR})
 
 BISON_TARGET(parser parser.y ${CMAKE_CURRENT_BINARY_DIR}/parser.c)
 FLEX_TARGET(scanner lex.l  ${CMAKE_CURRENT_BINARY_DIR}/lexer.c)
 ADD_FLEX_BISON_DEPENDENCY(scanner parser)
 
-add_executable(d16
-  main.c instruction.c label.c
+add_library(d16-lib STATIC
+  assembler.c instruction.c label.c
   ${BISON_parser_OUTPUTS}
   ${FLEX_scanner_OUTPUTS})
-target_link_libraries(d16 "${Glib_LIBRARY}")
+ 
+target_include_directories(d16-lib PUBLIC
+  ${PROJECT_BINARY_DIR} ${PROJECT_SOURCE_DIR}
+  ${Glib_INCLUDE_DIR} ${GlibConfig_INCLUDE_DIR}
+)
+
+target_link_libraries(d16-lib
+  ${Glib_LIBRARY}
+)
