+++ assembler/d16-lib/assembler.h
@@ -5,12 +5,12 @@
 //  Created by Michael Nolan on 6/17/16.
 //  Copyright © 2016 Michael Nolan. All rights reserved.
 //
+#include <glib.h>
+#include "instruction.h"
 
 #ifndef main_h
 #define main_h
-#include <glib.h>
-#include "instruction.h"
-void process_list(struct _GList*);
+void process_list(struct _GList* list, FILE* output_file);
 void print_elem(void* element, void* data);
 void free_elem(void* element);
 void assemble_instruction(Instruction* i, uint16_t** data);
