+++ assembler/d16-main/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_executable(d16 main.c)
+target_link_libraries(d16 d16-lib)
